\name{addHandlerChanged<-.GTable}
\alias{addHandlerChanged<-.GTable}
\title{addHandlerChanged method}
\arguments{
  \item{}{}
}
\description{
  The change handler for \code{GTable} is called when a row
  is activated, such as when it is double clicked. A single
  click is for adjusting the selection.
}

