\name{editable}
\alias{editable}
\alias{editable.default}
\alias{editable<-}
\alias{editable<-.default}
\title{Controls whether widget is editable or not}
\usage{
  editable(obj, j)

  editable.default(obj, j)

  editable(obj) <- value
}
\arguments{
  \item{obj}{object}

  \item{value}{logical. Set editable state.}
}
\description{
  Some widgets may be editable. If possible, the setter
  method can be used to toggle the state. This method
  indicates the state.

  Basic S3 method

  Set whether an object can be edited

  Basic S3 method for editable
}

