\name{font}
\alias{font}
\alias{font.default}
\alias{font<-}
\alias{font<-.default}
\title{Returns font specification for widget, if available}
\usage{
  font(obj)

  font.default(obj)

  font(obj) <- value
}
\arguments{
  \item{obj}{object}

  \item{value}{The font specification ... XXX ...}
}
\description{
  Returns font specification for widget, if available

  Basic S3 method for font

  Set font for a widget

  Basic S3 method for setting font
}

