% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_index_in_list}
\alias{get_index_in_list}
\title{get index of element of list}
\usage{
get_index_in_list(lst, ele)
}
\arguments{
\item{lst}{a list to search through}

\item{ele}{element of list}
}
\value{
returns index of element or integer(0)
}
\description{
Like match, but works with a list
}

