\name{getWithDefault}
\alias{getWithDefault}
\title{Return x unless NULL, NA, length 0, ..., in which case we give default}
\usage{
  getWithDefault(x, default)
}
\arguments{
  \item{x}{value to return or its default}

  \item{default}{default value}
}
\value{
  x or default
}
\description{
  Return x unless NULL, NA, length 0, ..., in which case we
  give default
}

