\name{.gfilebrowse}
\alias{.gfilebrowse}
\title{generic for toolkit dispatch}
\usage{
  .gfilebrowse(toolkit, text = "Select a file...", type =
  c("open", "save", "selectdir"), initial.filename = NULL,
  filter = list(), quote = TRUE, handler = NULL, action =
  NULL, container = NULL, ...)
}
\description{
  generic for toolkit dispatch
}

