\name{gseparator}
\alias{.gseparator}
\alias{gseparator}
\title{constructor providing a widget for displaying a line in a GUI}
\usage{
  gseparator(horizontal = TRUE, container = NULL, ...,
  toolkit = guiToolkit())

  .gseparator(toolkit, horizontal = TRUE, container = NULL,
  ...)
}
\description{
  constructor providing a widget for displaying a line in a
  GUI

  generic for toolkit dispatch
}

