\name{gspinbutton}
\alias{.gspinbutton}
\alias{gspinbutton}
\title{Spinbutton constructor}
\usage{
  gspinbutton(from = 0, to = 10, by = 1, length.out = NULL,
  along.with = NULL, value = from, digits = 0, handler =
  NULL, action = NULL, container = NULL, ..., toolkit =
  guiToolkit())

  .gspinbutton(toolkit, from = 0, to = 10, by = 1, value =
  from, digits = 0, handler = NULL, action = NULL,
  container = NULL, ...)
}
\description{
  Spinbutton constructor

  generic for toolkit dispatch
}

