\name{gstackwidget}
\alias{.gstackwidget}
\alias{gstackwidget}
\title{Constructor for a stack of widgets}
\usage{
  gstackwidget(container = NULL, ..., toolkit =
  guiToolkit())

  .gstackwidget(toolkit, container = NULL, ...)
}
\description{
  This widget is like a notebook -- it holds a stack of
  pages, but does not provide the tabs to work with. Most
  methods are inherited from gnotebook's.

  generic for toolkit dispatch
}

