\name{gstatusbar}
\alias{.gstatusbar}
\alias{gstatusbar}
\title{constructor to add a status bar to main window}
\usage{
  gstatusbar(text = "", container = NULL, ..., toolkit =
  guiToolkit())

  .gstatusbar(toolkit, text = "", container = NULL, ...)
}
\description{
  constructor to add a status bar to main window

  generic for toolkit dispatch
}

