\name{str.GComponent}
\alias{merge.list}
\alias{str.GComponent}
\title{str method for widgets}
\usage{
  str.GComponent(object, ...)

  merge.list(x, y, overwrite = TRUE)
}
\arguments{
  \item{x}{a list}

  \item{y}{a list}

  \item{overwrite}{logical should we overright values in x}
}
\description{
  str method for widgets

  merge two lists
}

