% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gdfnotebook.R
\name{.gdfnotebook.default}
\alias{.gdfnotebook.default}
\title{Toolkit constructor}
\usage{
\method{.gdfnotebook}{default}(toolkit, items, container = NULL, ...)
}
\arguments{
\item{toolkit}{toolkit}

\item{items}{data frame for initial page, when given}

\item{container}{parent container}

\item{...}{passed to \code{add} method of parent container}
}
\description{
Toolkit constructor
}

