% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{is_Windows}
\alias{is_Windows}
\title{Return logical indicating if we are on a Windows machine}
\usage{
is_Windows()
}
\value{
logical
}
\description{
Return logical indicating if we are on a Windows machine
}

