% Generated by roxygen2 (4.0.1): do not edit by hand
\name{observer}
\alias{observer}
\title{constructor for handler object}
\usage{
observer(receiver, handler, action = NULL)
}
\arguments{
\item{receiver}{object receiving event}

\item{handler}{function to call}

\item{action}{used to parametrize handler call
not exported, call using :::}
}
\description{
constructor for handler object
}

