% Please edit documentation in R/methods.R
\name{redo}
\alias{redo}
\alias{redo.GComponent}
\title{Redo past action.}
\usage{
redo(obj, ...)

\method{redo}{GComponent}(obj, ...)
}
\arguments{
\item{obj}{object to redo}

\item{...}{ignored}
}
\description{
Some widgets support redo actions

S3 method.
}

