\name{registered_packages}
\alias{registered_packages}
\title{##  toolkit class for rJava
##  toolkit class for SJava
##  toolkit class for  tcltk
##  toolkit class for  RwxWidgets
all packages that are registered with gWidgets2. Used if guiToolkit not specified}
\description{
  ## toolkit class for rJava ## toolkit class for SJava ##
  toolkit class for tcltk ## toolkit class for RwxWidgets
  all packages that are registered with gWidgets2. Used if
  guiToolkit not specified
}

