\name{short_summary}
\alias{short_summary}
\alias{short_summary.character}
\alias{short_summary.data.frame}
\alias{short_summary.default}
\alias{short_summary.function}
\alias{short_summary.list}
\alias{short_summary.lm}
\alias{short_summary.logical}
\alias{short_summary.matrix}
\alias{short_summary.numeric}
\title{Provide a short summary for an object}
\usage{
  short_summary(x)

  short_summary.default(x)

  short_summary.numeric(x)

  short_summary.character(x)

  short_summary.logical(x)

  short_summary.data.frame(x)

  short_summary.matrix(x)

  short_summary.list(x)

  short_summary.lm(x)

  short_summary.function(x)
}
\arguments{
  \item{x}{object}
}
\description{
  Provide a short summary for an object

  method for generic

  method for generic

  method for generic

  method for generic

  method for generic

  method for generic

  method for generic

  method for generic

  method for generic
}

