\name{tooltip}
\alias{isExtant.default}
\alias{tooltip}
\alias{tooltip<-}
\alias{tooltip<-.default}
\title{Get a tooltip for the widget}
\usage{
  tooltip(obj) <- value
}
\arguments{
  \item{obj}{object}

  \item{obj}{object}

  \item{value}{character tooltip value}
}
\description{
  Get a tooltip for the widget

  Basic S3 method for tooltip<-

  Set a tooltip for the widget

  Basic S3 method for tooltip<-

  Basic S3 method for isExtant
}

