% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{undo}
\alias{undo}
\alias{undo.GComponent}
\title{Undo past action.}
\usage{
undo(obj, ...)

\method{undo}{GComponent}(obj, ...)
}
\arguments{
\item{obj}{object to call undo on}

\item{...}{ignored}
}
\description{
Some widgets support undo actions. See reference class method \code{can_undo} as well.

S3 method.
}

