\name{WSWatcherModel}
\alias{WSWatcherModel}
\title{A class for monitoring workspacke}
\description{
  A reference class to create a model that monitors the
  global workspace. The class has methods
  \code{update_state}, \code{start_timer},
  \code{stop_timer} and the "getting" methods
  \code{get_by_class}, \code{get_by_function} (filter),
  \code{get_changes}
}

