% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{XXX}
\alias{XXX}
\title{Functions to message something needs doing. Easy to search for}
\usage{
XXX(msg)
}
\arguments{
\item{msg}{optional message to emit}
}
\description{
Functions to message something needs doing. Easy to search for
}

