% Generated by roxygen2 (4.0.1): do not edit by hand
\name{editable}
\alias{editable}
\alias{editable.default}
\alias{editable<-}
\alias{editable<-.default}
\title{Controls whether widget is editable or not}
\usage{
editable(obj, i)

\method{editable}{default}(obj, i)

editable(obj, i) <- value

editable(obj, i) <- value
}
\arguments{
\item{obj}{object}

\item{i}{index to apply to, when applicable}

\item{value}{logical. Set editable state.}
}
\description{
Some widgets may be editable. If possible, the setter method can
be used to toggle the state. This method indicates the state.

Basic S3 method

Set whether an object can be edited

Basic S3 method for editable
}

