% Generated by roxygen2 (4.0.1): do not edit by hand
\name{font<-.GText}
\alias{font}
\alias{font.default}
\alias{font<-}
\alias{font<-.GText}
\alias{font<-.default}
\title{Set font for gtext object}
\usage{
\method{font}{GText} (obj) <- value

font(obj)

\method{font}{default}(obj)

font(obj) <- value

font(obj) <- value
}
\arguments{
\item{obj}{object}

\item{value}{The font specification is given in terms of a named
vector or list where the names indicate a font attribute and the
value a reasonable choice:

\describe{

\item{weight}{c("light", "normal", "medium", "bold", "heavy")}

\item{style}{c("normal", "oblique", "italic")}

\item{family}{c("sans", "helvetica", "times", "monospace")}

\item{size}{an integer, say c(6,8,10,11,12,14,16,18,20, 24,36,72)}

\item{color (or foreground)}{One of colors()}

\item{background}{One of colors()}

\item{scale}{c("xx-large", "x-large",  "large" ,   "medium",   "small",    "x-small",  "xx-small")}

}

These are from Gtk's font specs, which though fairly standard, may
not be totally supported in the other toolkits.}
}
\description{
The \code{font} assignment method is used to change the font of
the currently selected text.

Returns font specification for widget, if available

Basic S3 method for font

Set font for a widget

Basic S3 method for setting font
}

