\name{[<-.GComboBox}
\alias{[<-.GComboBox}
\title{replacement method for combobox selection items}
\arguments{
  \item{x}{object}

  \item{i}{ignored}

  \item{j}{ignored}

  \item{...}{ignored}

  \item{value}{new items for selection}
}
\description{
  Ensure that value is a data frame. One can pass a vector
  or a one-column data frame to inidicate the possible
  values for selection, a second column is used for an
  icons (if possible), a third for a tooltip (if possible).
  get_from_templace for guiWidget
}

