% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicInterface.R, R/S3-methods.R, R/gdf.R,
%   R/gedit.R, R/gframe.R
\name{[.GDefaultWidget}
\alias{[.GDefaultWidget}
\alias{length.GComponent}
\alias{length<-.GComponent}
\alias{dim.GComponent}
\alias{names.GComponent}
\alias{names<-.GComponent}
\alias{dimnames.GComponent}
\alias{dimnames<-.GComponent}
\alias{[.GComponent}
\alias{[.GContainer}
\alias{[<-.GComponent}
\alias{update.GComponent}
\alias{str.GComponent}
\alias{[.GDf}
\alias{[.GEdit}
\alias{names<-.GFrame}
\title{Return items}
\usage{
\method{[}{GDefaultWidget}(x, i, j, ...)

\method{length}{GComponent}(x)

\method{length}{GComponent} (x) <- value

\method{dim}{GComponent}(x)

\method{names}{GComponent}(x)

\method{names}{GComponent} (x) <- value

\method{dimnames}{GComponent}(x)

\method{dimnames}{GComponent} (x) <- value

\method{[}{GComponent}(x, i, j, ..., drop = TRUE)

\method{[}{GContainer}(x, i, j, ..., drop = TRUE)

\method{[}{GComponent} (x, i, j, ...) <- value

\method{update}{GComponent}(object, ...)

\method{str}{GComponent}(object, ...)

\method{[}{GDf}(x, i, j, ..., drop = TRUE)

\method{[}{GEdit}(x, i, j, ..., drop = TRUE)

\method{names}{GFrame} (x) <- value
}
\arguments{
\item{x}{component}

\item{i}{index or row index if applicable}

\item{j}{column index if applicable}

\item{...}{dots argument}

\item{value}{value to assign}

\item{drop}{logical. Does return value get "dropped" down to something easier?}

\item{object}{object to update}
}
\value{
length of object
}
\description{
Names are used in many different contexts.

We use the extraction operator, \code{[}, typically to refer to
the underlying items from which a selection can be made. As well,
we overload this to containers to refer to the child components.

The update method will ca
use a widget to recompute itself, if it is necessary.

The current items for a \code{gdf} object are both the visible and
non-visible items. To retrieve just the currently visible items,
use the idiom \code{obj[visible(obj), ]}.

The underlying widget may allow autocompletion, if this is the
case then this method is used to set the list of candidates.
}
