\name{merge.list}
\alias{merge.list}
\alias{str.GComponent}
\title{merge two lists}
\usage{
  merge.list(x, y, ...)

  str.GComponent(object, ...)
}
\arguments{
  \item{x}{a list}

  \item{y}{a list}

  \item{overwrite}{logical should we overright values in x}
}
\description{
  merge two lists

  str method for widgets
}

