\name{addSpring}
\alias{addSpring}
\alias{addSpring.GContainer}
\title{Add a spring to a container objects}
\usage{
  addSpring(obj)

  addSpring.GContainer(obj)
}
\arguments{
  \item{obj}{GContainer object}
}
\value{
  NULL
}
\description{
  Add a spring to a container objects

  Add spring to GContainer class
}

