\name{addSpring}
\alias{addSpace}
\alias{addSpace.GContainer}
\alias{addSpring}
\alias{addSpring.GContainer}
\title{Add a spring to a container objects}
\usage{
  addSpring(obj)

  \method{addSpring}{GContainer} (obj)

  addSpace(obj, value)

  \method{addSpace}{GContainer} (obj, value)
}
\arguments{
  \item{obj}{GContainer object}

  \item{obj}{GContainer object}

  \item{value}{space in pixels to add}
}
\value{
  NULL

  NULL
}
\description{
  Add a spring to a container objects

  Add spring to GContainer class

  Add a spring to a container objects

  Add space to GContainer class
}

