% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{size}
\alias{size}
\alias{size.default}
\alias{size<-}
\alias{size<-.default}
\title{Return size (width and height) of widget}
\usage{
size(obj)

\method{size}{default}(obj)

size(obj) <- value

size(obj) <- value
}
\arguments{
\item{obj}{object}

\item{value}{size in pixels}
}
\description{
The size is specified in pixels (integers). Some toolkits allow -1 as a default, but not all.
}
