% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{tooltip}
\alias{tooltip}
\alias{tooltip.default}
\alias{tooltip<-}
\alias{tooltip<-.default}
\title{Get a tooltip for the widget}
\usage{
tooltip(obj)

\method{tooltip}{default}(obj)

tooltip(obj) <- value

tooltip(obj) <- value
}
\arguments{
\item{obj}{object}

\item{value}{character tooltip value}
}
\description{
Get a tooltip for the widget

Basic S3 method for tooltip<-

Set a tooltip for the widget

Basic S3 method for tooltip<-
}

