\name{undo}
\alias{undo}
\alias{undo.GComponent}
\title{Undo past action.}
\usage{
  undo(obj, ...)

  \method{undo}{GComponent} (obj, ...)
}
\arguments{
  \item{obj}{object to call undo on}

  \item{...}{ignored}
}
\description{
  Some widgets support undo actions. See reference class
  method \code{can_undo} as well.

  S3 method.
}

